refactor(emit): split class recovery helpers
